+++ src/pages/Main/Main.jsx
@@ -1,12 +1,14 @@
 import { styled } from 'styled-components';
 import CafeList from '../Cafe/CafeList';
-import KakaoMaps from '../../hooks/KakaoMaps/kakaoMaps';
+import KakaoMaps from '../../pages/KakaoMaps/kakaoMaps';
+import Map from '../KakaoMaps/Map';
 
 const Main = () => {
   return (
     <MainBody>
       <MapBox>
         <KakaoMaps />
+        {/* <Map /> */}
       </MapBox>
       <CafeListBox>
         <CafeList />
